refactor: clean up imports and move code out of __init__.py files
- Move tmux navigation code from __init__.py to tmux/navigation.py
- Move wezterm navigation code from __init__.py to wezterm/navigation.py
- Replace all `from .foo import bar as baz` renames with package
  imports (`from . import foo`) and dotted access (`foo.bar`)
- Package __init__.py files now just export submodules
- Remove unused barrel reexports from inbox/__init__.py
- Remove underscore-rename of openclaw find_session_path import
- Un-defer imports where no circular dependency exists

migrations/__init__.py kept as-is: discover_migrations() uses
__path__/__package__ which only work in __init__.py.
