+++ src/lemonaid/codex/__init__.py
@@ -1,1 +1,3 @@
 """Lemonaid Codex CLI integration."""
+
+from . import cli, watcher  # noqa: F401
