+++ src/lemonaid/openclaw/__init__.py
@@ -6,3 +6,5 @@ OpenClaw is an open-source personal AI assistant. Sessions are stored at:
 Session index is at:
 ~/.openclaw/agents/<agentId>/sessions/sessions.json
 """
+
+from . import cli, utils, watcher  # noqa: F401
