+++ src/lemonaid/opencode/__init__.py
@@ -3,3 +3,5 @@
 OpenCode stores session state in SQLite at:
 ~/.local/share/opencode/opencode.db
 """
+
+from . import cli, watcher  # noqa: F401
